More templates.

+++ templates/scripts/eject.sh
@@ -2,7 +2,12 @@
 echo "Fetching stuff from inside the container."
 
 IMAGE_NAME=s_vue_2nd
-CONTAINER_NAME=$(docker ps | grep $IMAGE_NAME | sed 's/ \+ /\t/g' | cut -d$'\t' -f7)
+RUNNER=./$IMAGE_NAME.sh
+if [ ! -f $RUNNER ]; then exit 1; fi
+CONTAINER_NAME=${IMAGE_NAME}_tmp
+CONTAINER_NAME=$CONTAINER_NAME NO_TTY=1 $RUNNER eval "sleep 30;" &
+
+sleep 3
 
 docker cp $CONTAINER_NAME:/$IMAGE_NAME/node_modules ./
 docker cp $CONTAINER_NAME:/$IMAGE_NAME/.eslintrc.js ./
